# projectT4T/storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "API Documentation",
        "version": "1.0.0"
    },
    "host": "localhost:8000",
    "basePath": "",
    "schemes": [
        "http",
        "https"
    ],
    "paths": {
...
        "/api/GetTrees": {
            "get": {
                "tags": [
                    "Trees"
                ],
                "summary": "Get Trees",
                "operationId": "GetTrees",
                "parameters": [
                    {
                        "name": "tree_name",
                        "in": "query",
                        "type": "string"
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "type": "integer"
                    },
                    {
                        "name": "offset",
                        "in": "query",
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/GetTreesDetail": {
            "get": {
                "tags": [
                    "Trees"
                ],
                "summary": "Get Trees Detail",
                "operationId": "GetTreesDetail",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/AddTrees": {
            "post": {
                "tags": [
                    "Trees"
                ],
                "summary": "Add Trees",
                "operationId": "AddTrees",
                "parameters": [
                    {
                        "name": "Parameters",
                        "in": "body",
                        "description": "Add Trees",
                        "required": true,
                        "schema": {
                            "properties": {
                                "tree_name": {
                                    "type": "string",
                                    "example": "Mangga"
                                },
                                "scientific_name": {
                                    "type": "string",
                                    "example": "Mangifera indica"
                                },
                                "common_name": {
                                    "type": "string",
                                    "example": "Spanish Cherry"
                                },
                                "short_information": {
                                    "type": "string",
                                    "example": "Mangifera indica is a large evergreen tree to 20 m tall with a dark green. umbrella-shaped crown"
                                },
                                "tree_category": {
                                    "type": "string",
                                    "example": "Pohon"
                                },
                                "english_name": {
                                    "type": "string",
                                    "example": "Nullable"
                                },
                                "description": {
                                    "type": "string",
                                    "example": "Nullable"
                                },
                                "product_list": {
                                    "type": "string",
                                    "example": "Nullable"
                                },
                                "estimate_income": {
                                    "type": "string",
                                    "example": "Nullable"
                                },
                                "co2_capture": {
                                    "type": "integer",
                                    "example": 0
                                },
                                "photo1": {
                                    "type": "string",
                                    "example": "Nullable"
                                },
                                "photo2": {
                                    "type": "string",
                                    "example": "Nullable"
                                }
                            }
                        },
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/UpdateTrees": {
            "post": {
                "tags": [
                    "Trees"
                ],
                "summary": "Update Trees",
                "operationId": "UpdateTrees",
                "parameters": [
                    {
                        "name": "Parameters",
                        "in": "body",
                        "description": "Update Trees",
                        "required": true,
                        "schema": {
                            "properties": {
                                "id": {
                                    "type": "integer",
                                    "example": 1
                                },
                                "tree_name": {
                                    "type": "string",
                                    "example": "Mangga"
                                },
                                "scientific_name": {
                                    "type": "string",
                                    "example": "Mangifera indica"
                                },
                                "common_name": {
                                    "type": "string",
                                    "example": "Spanish Cherry"
                                },
                                "short_information": {
                                    "type": "string",
                                    "example": "Mangifera indica is a large evergreen tree to 20 m tall with a dark green. umbrella-shaped crown"
                                },
                                "tree_category": {
                                    "type": "string",
                                    "example": "Pohon"
                                },
                                "english_name": {
                                    "type": "string",
                                    "example": "Nullable"
                                },
                                "description": {
                                    "type": "string",
                                    "example": "Nullable"
                                },
                                "product_list": {
                                    "type": "string",
                                    "example": "Nullable"
                                },
                                "estimate_income": {
                                    "type": "string",
                                    "example": "Nullable"
                                },
                                "co2_capture": {
                                    "type": "integer",
                                    "example": 0
                                },
                                "photo1": {
                                    "type": "string",
                                    "example": "Nullable"
                                },
                                "photo2": {
                                    "type": "string",
                                    "example": "Nullable"
                                }
                            }
                        },
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/DeleteTrees": {
            "post": {
                "tags": [
                    "Trees"
                ],
                "summary": "Delete Trees",
                "operationId": "DeleteTrees",
                "parameters": [
                    {
                        "name": "Parameters",
                        "in": "body",
                        "description": "Delete Trees",
                        "required": true,
                        "schema": {
                            "properties": {
                                "id": {
                                    "type": "integer",
                                    "example": 1
                                }
                            }
                        },
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
...
    },
...
    "securityDefinitions": {
        "apiAuth": {
            "type": "apiKey",
            "description": "Login with email and password to get the authentication token",
            "name": "Authorization",
            "in": "header"
        }
    }
}